ExUnit v1.10.3 ExUnit.TestModule

A struct that keeps information about the test module.

It is received by formatters and contains the following fields:

:name - the test module name

:state - the test error state (see ExUnit.state/0)

:tests - all tests in this module

Types

t()

Specs

t() :: %ExUnit.TestModule{
  name: module(),
  state: ExUnit.state(),
  tests: [ExUnit.Test.t()]
}