ExUnit.TestModule (ExUnit v1.15.0-dev)

A struct that keeps information about the test module.

It is received by formatters and contains the following fields:

:file - (since v1.11.0) the file of the test module

:name - the test module name

:state - the test error state (see ExUnit.state/0)

:tests - all tests in this module

Types

t()

@type t() :: %ExUnit.TestModule{
  file: binary(),
  name: module(),
  state: ExUnit.state(),
  tags: map(),
  tests: [ExUnit.Test.t()]
}